// guipmc-casuistica/src/app/casuistica/models/search-records/detailSearchRecordsDTO.ts
export class DetailSearchRecordsDTO {
  numNss: string;
  nomCompletoAsegurado: string;
  refRegistroPatronal: string;
  fecInicio: string;
  desConsecuencia: string;
  cveConsecuencia: number;
  desTipoRiesgo: string;
  cveTipoRiesgo: string;
  numDiasSubsidiados: string;
  porcentajeIncapacidad: string;
  fecFin: string;
  cveEstadoRegistro: string;
  desEstadoRegistro: string;
  numFolioMovtoOriginal: string;
  objectId: string;
  position: number;
  isChange: boolean;
  cveSituacionRegistro: number;
  desSituacionRegistro: string;
  cveOrigenArchivo: string;
  fecActualizacion: Date;
}